Fix duplicate detection in Solver::learn()
